server/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
langchain-text-splitters
langchain-pinecone
pinecone-client
langchain-voyageai
langchain-groq
langgraph
pydantic
python-dotenv